requirements.txt
absl-py==1.0.0
astunparse==1.6.3
autopep8==1.6.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cycler==0.11.0
easyocr==1.4.1
editdistance==0.6.0
fire==0.4.0
flatbuffers==2.0
fonttools==4.28.3
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
idna==3.3
imageio==2.9.0
importlib-metadata==4.8.2
imutils==0.5.4
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lmdb==1.3.0
Markdown==3.3.6
matplotlib==3.5.0
natsort==8.0.2
networkx==2.6.3
nltk==3.6.7
numpy==1.21.4
oauthlib==3.1.1
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pafy==0.5.5
Pillow==8.2.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyparsing==3.0.6
pytesseract==0.3.8
python-bidi==0.4.2
python-dateutil==2.8.2
PyWavelets==1.2.0
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.0
scipy==1.7.3
setuptools-scm==6.3.2
six==1.16.0
split-folders==0.4.3
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
tifffile==2021.11.2
toml==0.10.2
tomli==1.2.2
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wincertstore==0.2
wrapt==1.13.3
youtube-dl==2021.12.17
zipp==3.6.0
